:new: :art: working-groups: Document Design & UX magic (closes #296)
This commit adds clarity into the things we do as a Working Group and
also explains how a new-comer could participate with the Working Group.
The "How to get involved" section is sorted in order of difficulty,
where items on the top are easiest to complete and take the lowest
commitment. Items towards the bottom require some degree of commitment.

This is a first pass at documenting how we work, definitely room for
improvements and other iterations here. :grinning:

Closes #296.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>